requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
tabula-py==2.3.0
pdfplumber==0.7.1
camelot-py==0.10.1